Fixed the weird timer error thing.
I guess it was caused by the timer having no parent making it be destroyed after the event loop had gone. I think the event loop is making timers work so I guess it's no surprise that it would break. Anyway now the timer hasd the main root widget as a parent.